+++ client/src/components/App/App.js
@@ -1,0 +1,16 @@
+import React from "react";
+import "./App.css";
+import Form from "../Form/Form";
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <h1>Input Data</h1>
+      </header>
+      <Form />
+    </div>
+  );
+}
+
+export default App;
