+++ client/src/components/Form/Form.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Form() {
+  return (
+    <form>
+      <input type="text"></input>
+    </form>
+  );
+}
+
+export default Form;
